\name{pages-package}
\Rdversion{1.1}
\alias{pages-package}
\alias{pages}
\docType{package}
\title{
Peer group analysis using returns-based performance and risk measures
}
\description{
Creates formatted views of performance and risk information using PerformanceAnalytics and other packages. This is considered EXPERIMENTAL CODE and WILL NOT BE SUPPORTED.  Collaboration towards creating a supportable code base would be welcome, however.
}
\details{
\tabular{ll}{
Package: \tab pages\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-28\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Peter Carl
}

